Move CSV spec utility methods into a module
